working auth with google and making bottom home profile and chat and nfc
feat: Revise layout positioning in LoginScreen and implement navigation

This commit adjusts the positioning of the Google and icon images in the
LoginScreen component, as well as the rectangle background. The left
positions have been updated to better align with the design
requirements. Additionally, the width, height, and left-right margins of
the rectangle image have been modified to create a desired layout.

Moreover, the App component has been updated to include the necessary
dependencies for implementing navigation. The ClerkProvider and
NavigationContainer components have been added to wrap the content, and
the TabNavigation component has been introduced to handle navigation
logic.

These changes improve the visual alignment of elements and enable the
implementation of bottom navigation with